split validation / constants
