+++ src/components/Asset/Edit/_types.ts
@@ -1,10 +1,12 @@
+import { FileInfo } from '@oceanprotocol/lib'
+
 export interface MetadataEditForm {
   name: string
   description: string
   timeout: string
   price?: string
-  links?: string | any[]
-  files: string | any[]
+  files: FileInfo[]
+  links?: FileInfo[]
   author?: string
 }
